+++ pytest_telegram/plugin.py
@@ -46,6 +46,18 @@ def pytest_addoption(parser):
         default=None,
         help='Custom text which will added to telegram message'
     )
+    group.addoption(
+        '--telegram_disable_stickers',
+        action='store_true',
+        dest='telegram_disable_stickers',
+        help='Option for disable stickers'
+    )
+    group.addoption(
+        '--telegram_list_failed',
+        action='store_true',
+        dest='telegram_list_failed',
+        help='Option for list failed tests'
+    )
 
 
 @pytest.hookimpl(hookwrapper=True)
@@ -71,16 +83,42 @@ def pytest_terminal_summary(terminalreporter, exitstatus, config):
     success_sticker_id = config.option.success_sticker_id
     fail_sticker_id = config.option.fail_sticker_id
     report_url = f'\n{config.option.telegram_report_url}' if config.option.telegram_report_url else ''
-    custom_text = f'\n{config.option.telegram_custom_text}'.replace('\\n', '\n') if config.option.telegram_custom_text else ''
+    custom_text = f'\n{config.option.telegram_custom_text}'.replace('\\n',
+                                                                    '\n') if config.option.telegram_custom_text else ''
+    disable_stickers = config.option.telegram_disable_stickers
+    list_failed = config.option.telegram_list_failed
+    list_failed_amount = 3
 
+    failed_tests = ''
+    error_tests = ''
+    if list_failed and failed != 0:
+        failed_tests = '\nFailed tests:\n'
+
+        for failed_test in terminalreporter.stats.get('failed', [])[:list_failed_amount]:
+            failed_tests += f'{failed_test.nodeid}\n'
+
+        if failed > list_failed_amount:
+            failed_tests += '...'
+    if list_failed and error != 0:
+        error_tests = '\nError tests:\n'
+
+        for error_test in terminalreporter.stats.get('error', [])[:list_failed_amount]:
+            error_tests += f'{error_test.nodeid}\n'
+
+        if error > list_failed_amount:
+            error_tests += '...'
     final_results = 'Passed=%s Failed=%s Skipped=%s Error=%s XFailed=%s XPassed=%s' % (
         passed, failed, skipped, error, xfailed, xpassed)
 
     if failed == 0 and error == 0:
-        payload = {'chat_id': chat_id, 'sticker': success_sticker_id}
+        sticker_payload = {'chat_id': chat_id, 'sticker': success_sticker_id}
     else:
-        payload = {'chat_id': chat_id, 'sticker': fail_sticker_id}
+        sticker_payload = {'chat_id': chat_id, 'sticker': fail_sticker_id}
 
-    message_id = requests.post(f'{telegram_uri}/sendSticker', json=payload).json()['result']['message_id']
-    payload = {'chat_id': chat_id, 'text': f'{final_results}{custom_text}{report_url}', 'reply_to_message_id': message_id}
-    requests.post(f'{telegram_uri}/sendMessage', json=payload).json()
+    message_id = None
+    if not disable_stickers:
+        message_id = requests.post(f'{telegram_uri}/sendSticker', json=sticker_payload).json()['result']['message_id']
+    message_payload = {'chat_id': chat_id,
+                       'text': f'{final_results}{custom_text}{report_url}{failed_tests}{error_tests}',
+                       'reply_to_message_id': message_id}
+    requests.post(f'{telegram_uri}/sendMessage', json=message_payload).json()